* functions: change autostart to type reserved_bool_option
The "autostart" variable sometimes causes trouble because only the
"no autostart" form was working. Using "autostart" alone did in
fact disable the service. The reason is that it is a "reserved_option"
(a string) and only the default value "1" is accepted to indicate the
service is enabled.

So we create a new type "reserved_bool_option" for boolean variables
which are set to 1 when specified, and zero when declared with "no".
This fixes the problem, and "autostart" alone now has the expected
behaviour.