functions: make use of taskset when available to set the CPU mask
taskset is one of the ugliest tools ever invented. We had to go through
a number of hacks to work around its limitations, including forcing the
mask to hexadecimal representation using printf. We also had to first
build a mask of currently online CPUs to filter the mask based on this
so that taskset never fails with a sched_setaffinity() error.

When the utility is found, _CPUMASK is defined to point to a local
wrapper which does the conversion from cpumask to taskset format. When
it's not available, cpumask is still used.