preinit: only populate /dev when it's not a devtmpfs
It's pointless to pollute /dev with dummy entries when it's a tmpfs,
all the added ones do not work and some are even confusing. Thanks to
the new "td" command introduced in latest flxutils, we can detect if
a devtmpfs is mounted on /dev and refrain from doing the job. For older
versions, the missing command doesn't alter the status code, so we
preset it with a failure using the "eq" command, and if the failure
remains after "td" we continue to populate the FS as we used to.