preinit: stop calling depmod if there are no modules
We want to call it only when there's at least something in /lib/modules
otherwise it complains loudly at boot. The possibilities for the check
are quite limited so we have to call some shell on /lib/modules and to
use shift to detect the presence (or absence) of at least one entry.