functions: order the list of modules in best_effort_rmmod
The modprobe -r implementation that comes with kmod doesn't automatically
sort the modules according to their removal order like module-init-tools
used to. It happens that our best_effort_rmmod function used to provide
the list from the root to the leaves regarding the dependencies, which
makes the modules removal fail.

It would seem obvious that the list needs to be reversed but that doesn't
work because sometimes some modules do not list all of their dependencies
and can still be misplaced (eg: nf_conntrack_ipv4 when nf_nat and
iptable_nat are both loaded). However there's something always known with
modules, which is that they appear with the lastest ones at the top of the
list. So we know that we have to remove them respecting the list order in
order to always solve the dependency issue. It also has the benefit of
calling modprobe -r with a single occurence of each module.