dt-modules: new script to load modules for device-tree nodes
Some ARM platforms require loading certain modules to access controllers
that are enumerated in the device tree and without which nothing else
can be detected (e.g. PCIe controllers, MMC/SD controllers, etc). The
new dt-modules script scans the device tree nodes in /sys looking for
nodes that are supported by some of the modules, and loads them. It may
for example allow to find the flash device, a hard disk or an ethernet
controller. It also supports simply listing aliases or the modules.