network: replace "tunnel_type" and "master_type" with a more generic "type"
In order to ease addition of new types, let's rename "tunnel_type" to
"type" and make "master_type" also alias to it since both are exclusive.
For now "type" accepts "bonding", "tun" and "tap", and both "tunnel_type"
and "master_type" are just an alias that forces "type". "master_type" was
already not documented.

Previously, having slaves declared was sufficient to setup a bond. Now
we continue to do that but this implies setting the type to "bonding"
when not set while having slaves, which is cleaner as it maintains some
consistency between types and arguments.