network: add support for setting up wireguard interfaces
These ones are declared of type "wg". They will be created with the
listen port, fwmark, and private_key. The peers will need to be
configured separately. If we fail to bind a wg interface (e.g. port
already in use), we destroy it otherwise it's left in a state that
the user can neither start nor stop. A wg interface that was manually
set down still needs to be removed when stopping, so we add a check
for this as well.