network: remap "bridge" to "slave"
The "bridge" and "slave" options to exactly the same except that they
imply a different type of device. "bridge" implicitly creates a bridge
while "slave" implicitly creates a bond. Here we remap "bridge" to
"slave" so that we can always use "slave" to figure the list of slave
devices. Both continue to preset the type though, for backwards
compatibility.