network: add support for "ip rule" and "ip6 rule" for policy routing
These rules are applied on the inbound traffic of the interface. They
always filter on the interface name with the "iif" parameter, except
for rules in the unnamed section. Loopback should see outgoing traffic
here, even for sockets bound to an interface (check using oif for this).
The rules are carefully removed when stopping.