network: add support for LED triggers
Some boards come with interfaces having no LED, and instead they rely
on software-controllable LEDs. While we've implemented the alix-leds
project in the past to deal with this, modern kernels support LED
triggers for network interface link events.

This patch makes use of this mechanism. It provides a new "led" option
in the network section, which designates the LED name. Once up, the
interface will be attached there so that the link status is reflected
on the LED. There's one caveat however: turning the interface down will
automatically unregister it, which means that bridges and bonds turn
this off. For this reason, there's an explicit search there when attaching
a slave for a possible "led" in the interface being attached, to turn it
on again.