system: improve console auto-configuration
/dev/console is convenient for use as an always-working console but it
suffers from some shortcomings such as not supporting process group
ownership. We can do better thanks to /proc/consoles and /sys. In the
/proc/consoles file we have all consoles passed on the cmdline, with
their state (enabled or not) and the device number. The name might not
always be correct on some archis, but the device number is. In addition,
when seeing "tty0" we know that instead it's a list of VTs, and that tty0
itself must not be used.

Thus we add a new config option, "vt_concoles", which sets the number of
desired virtual terminals. We make it default to 4 when not set if tty0
is found in /proc/consoles, unless the value was forced to zero. This way
we automatically create VTs when they are requested by the boot loader.

In addition, all other real consoles are created and enabled as well.
The inittab file is updated with all this, with older entries trimmed to
avoid duplicates, and init is reloaded with the new values.