system: support activating extra consoles at the service level
In addition to autodetected consoles, we can now activate extra consoles
using the extra_consoles variable. It's multi-valued so it can get multiple
consoles either on one or multiple lines, all are cumulated. Since they're
all processed in order, it's even possible to redefine an already defined
console. The syntax for each of them is:

     device[:[speed][:[term]]]

The default speed will continue to be "-s" (don't change) if unspecified
or empty. The default term will remain "linux". But for example it's
possible to change the ttyS0 speed even if it's the console:

    extra_consoles ttyS0:115200

or to just change the terminal type:

    extra_consoles ttyS0::vt320

Since these consoles are processed just like the default ones, passing
any "tty[0-9]" will be sufficient to immediately enable the VTs with
the configured number of VTs or 4 by default.

It can make sense to have:

    extra_consoles tty0 ttyUSB0

in the defaults of a headless system to which connections may happen once
in a while by bringing a monitor or by connecting a USB adapter for example.